admin-api/requirements.txt
Flask
flask-redis
gunicorn
kubernetes
